+++ seogi/src/main/java/com/example/seogi/di/annotation/FieldInject.kt
@@ -1,4 +1,4 @@
-package com.example.seogi.di
+package com.example.seogi.di.annotation
 
 @Target(AnnotationTarget.CONSTRUCTOR, AnnotationTarget.PROPERTY)
 @Retention(AnnotationRetention.RUNTIME)
