+++ seogi/src/main/java/com/example/seogi/di/annotation/OnDisk.kt
@@ -1,4 +1,4 @@
-package com.example.seogi.di
+package com.example.seogi.di.annotation
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.VALUE_PARAMETER, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY)
