+++ seogi/src/main/java/com/example/seogi/di/annotation/Qualifier.kt
@@ -1,4 +1,4 @@
-package com.example.seogi.di
+package com.example.seogi.di.annotation
 
 @Target(AnnotationTarget.ANNOTATION_CLASS, AnnotationTarget.PROPERTY, AnnotationTarget.FIELD)
 @Retention(AnnotationRetention.RUNTIME)
